feat: Station 지하철 역 생성시 이름에 대한 예외 검사 기능 구현
- 지하철 역인 Station 객체는 이름이 null, 공백이 아닌 2글자 이상이다.
- 예외인 경우, StationNameException 커스텀 예외를 호출한다.
- 클래스 위치를 entity 패키지 내부로 이동시킨다.
