+++ src/main/java/subway/domain/StationRepository.java
@@ -1,5 +1,7 @@
 package subway.domain;
 
+import subway.domain.entity.Station;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
